fix: Prevent broken/strange checklist reordering
Updates the checklist reorder moves to filter and just consider those which are visually re-orderable within the UI, since odd behaviour could occur when the
fundemental task data had checked items between unchecked items while the "Move checked items to the bottom" setting was enabled.
This could occur from toggling the "Move checked items to the bottom" option, or by attempting to move a task item into/below the "Checked items" list.

This also adds a check when moving an item down, since the move handling would consider the "Checked items" as an element when emitting position changes.

Related to #59